lots of refactors and fixes
