+++ src/css/main.css
@@ -98,6 +98,9 @@ body[data-theme='dark'] {
     /* identical to box height, or 122% */
     text-transform: uppercase;
     color: var(--nav-link-color);
+    display: flex;
+    align-items: center;
+    justify-content: center;
 }
 
 .nav-link-collapsed,
@@ -140,6 +143,7 @@ body[data-theme='dark'] {
     vertical-align: middle;
     line-height: 1;
     margin: 0.2rem 0 0 0.5rem;
+    fill: none;
 }
 
 
@@ -763,28 +767,28 @@ button.close {
     position: absolute;
 }
 
-  .wrapper .unmoderated {
+.wrapper .unmoderated {
     position: absolute;
     left: 50%;
     top: 60%;
     transform: translate(-50%, -50%);
     width: 58px;
     height: 30px;
-  }
+}
 
-  .wrapper-toggle-unmoderated {
+.wrapper-toggle-unmoderated {
     position: relative;
     width: 80px;
     height: 15px;
 }
 
-  .wrapper .switch .unmoderated-slider {
+.wrapper .switch .unmoderated-slider {
     border: 1px solid #828282;
     background-color: transparent;
-  }
+}
 
-  .wrapper .switch .unmoderated-slider:before,
-  .wrapper .switch .unmoderated-slider:after {
+.wrapper .switch .unmoderated-slider:before,
+.wrapper .switch .unmoderated-slider:after {
     position: absolute;
     content: '';
     height: 24px;
@@ -794,20 +798,21 @@ button.close {
     border-radius: 24px;
     background-color: #999999ff;
     opacity: 1;
-  }
+}
 
-  .wrapper .switch input:checked + .unmoderated-slider:before {
+.wrapper .switch input:checked+.unmoderated-slider:before {
     transform: translateX(24px);
     background-color: #c81f1f;
-  }
-  .wrapper .switch input:checked + .unmoderated-slider:after {
+}
+
+.wrapper .switch input:checked+.unmoderated-slider:after {
     transform: translateX(24px);
     background-color: #c81f1f;
-  }
+}
 
-  .wrapper .switch input:checked + .unmoderated-slider {
+.wrapper .switch input:checked+.unmoderated-slider {
     background-color: #ff4f4f;
-  }
+}
 
 [data-theme='dark'] #browse-more {
     color: #dde3e9;
@@ -932,16 +937,6 @@ button.close {
 }
 
 
-/* .tweeter-meme{
-    background: url("../images/tweet_meme/meme-template.jpg");
-    background-size:"cover";
-    background-position:"center";
-    background-repeat: no-repeat;
-    width: 1200px;
-    height: 676px;
-    position: relative;
-} */
-
 .site-search {
 
     font-size: 48px;
@@ -959,4 +954,10 @@ button.close {
     position: absolute;
     left: 600px;
     top: 260px;
+}
+
+li.nav-item {
+    display: flex;
+    align-items: center;
+    justify-content: center;
 }
